SYCL: split lightweight detail headers out of umbrella paths
Split general-purpose utility umbrellas into narrow internal headers
so users that only need one helper stop paying for unrelated machinery.
These change save around 30ms when building `sycl/ext/oneapi/free_function_queries.hpp`,
coming from removing 17 header includes (and their transitive dependencies)
that were not actually needed for the building of `group.hpp`

* detail/assert.hpp: extracted __SYCL_ASSERT macro from common.hpp
  into its own minimal header; common.hpp now includes it.

* detail/loop.hpp: extracted detail::loop / loop_impl from helpers.hpp
  into a standalone header; retargeted accessor.hpp, group_algorithm.hpp,
  detail/builtins/builtins.hpp, and source/builtins/host_helper_macros.hpp
  to include the narrow helper directly.

* detail/nd_loop.hpp: extracted NDLoop, NDLoopIterateImpl, and
  InitializedVal from common.hpp; rewired cg_types.hpp and group.hpp
  to include nd_loop.hpp rather than the heavier common.hpp.

* detail/device_info_types.hpp: moved uuid_type / luid_type out of the
  broad type_traits.hpp into a dedicated device-info header; included
  that header from info/info_desc.hpp and relaxed the runtime check in
  device_impl.hpp to size + trivially-copyable requirements so the move
  stays source-compatible.

* group.hpp: replaced common.hpp / generic_type_traits.hpp /
  type_traits.hpp / item.hpp with the new narrow headers; added a
  private convertToOpenCLGroupAsyncCopyPtr helper that inlines the
  OpenCL pointer-conversion logic without pulling in the full generic
  conversion machinery.

* detail/async_work_group_copy_ptr.hpp: new narrow header providing
  async_copy_elem_type<T> and convertToOpenCLGroupAsyncCopyPtr.
  Dependencies are access/access.hpp, fwd/half.hpp, fwd/multi_ptr.hpp,
  <stdint.h>, <cstddef>, <type_traits> — all already required by
  any async_work_group_copy caller, so zero transitive cost is added.
  Uses std::make_signed_t / std::make_unsigned_t instead of hand-rolled
  fixed-width alias chains.

* detail/type_traits/bool_traits.hpp: new narrow header providing
  is_scalar_bool, is_vector_bool, is_bool, change_base_type_t.
  Depends only on vec_marray_traits.hpp + <type_traits>, so it does
  not pull in the heavier type_traits.hpp chain.  type_traits.hpp
  includes it and removes its own duplicate definitions, so existing
  callers are unaffected.

* group.hpp: remove inline group_async_copy_opencl_type family and
  convertToOpenCLGroupAsyncCopyPtr; include the new header instead.
  Drop now-unnecessary fwd/half.hpp, <cstddef>, and bfloat16 forward
  declaration (all moved into the new header).

* nd_item.hpp: replace #include <generic_type_traits.hpp> (which pulled
  in aliases.hpp, bit_cast.hpp, limits) with the new narrow header;
  replace ConvertToOpenCLType_t + DestT(ptr.get()) pattern with
  convertToOpenCLGroupAsyncCopyPtr(ptr) at all four call sites.

* test/include_deps/deps_known.sh: add sed rule for the
  unified-runtime/ subdirectory so ur_api.h and ur_api_funcs.def are
  stripped to bare filenames rather than emitting absolute build paths.

* test/include_deps/*.cpp: regenerated all golden files to reflect the
  updated include graphs and the deps_known.sh fix.
