+++ sycl/include/sycl/detail/builtins/builtins.hpp
@@ -64,8 +64,7 @@
 #pragma once
 
 #include <sycl/detail/fwd/multi_ptr.hpp>
-#include <sycl/detail/generic_type_traits.hpp>
-#include <sycl/detail/helpers.hpp>
+#include <sycl/detail/loop.hpp>
 #include <sycl/detail/type_traits.hpp>
 #include <sycl/detail/type_traits/vec_marray_traits.hpp>
 #include <sycl/half_type.hpp>
