+++ sycl/include/sycl/detail/cg_types.hpp
@@ -9,20 +9,20 @@
 #pragma once
 
 #include <sycl/detail/array.hpp>               // for array
-#include <sycl/detail/common.hpp>              // for InitializedVal, NDLoop
 #include <sycl/detail/helpers.hpp>             // for Builder
 #include <sycl/detail/host_profiling_info.hpp> // for HostProfilingInfo
 #include <sycl/detail/item_base.hpp>           // for id
 #include <sycl/detail/kernel_desc.hpp>         // for kernel_param_kind_t
+#include <sycl/detail/nd_loop.hpp>             // for InitializedVal, NDLoop
 #include <sycl/exception.hpp>
-#include <sycl/group.hpp>                      // for group
-#include <sycl/h_item.hpp>                     // for h_item
-#include <sycl/id.hpp>                         // for id
-#include <sycl/item.hpp>                       // for item
-#include <sycl/kernel_handler.hpp>             // for kernel_handler
-#include <sycl/nd_item.hpp>                    // for nd_item
-#include <sycl/nd_range.hpp>                   // for nd_range
-#include <sycl/range.hpp>                      // for range, operator*
+#include <sycl/group.hpp>          // for group
+#include <sycl/h_item.hpp>         // for h_item
+#include <sycl/id.hpp>             // for id
+#include <sycl/item.hpp>           // for item
+#include <sycl/kernel_handler.hpp> // for kernel_handler
+#include <sycl/nd_item.hpp>        // for nd_item
+#include <sycl/nd_range.hpp>       // for nd_range
+#include <sycl/range.hpp>          // for range, operator*
 
 #include <functional>  // for function
 #include <stddef.h>    // for size_t
