+++ sycl/include/sycl/group.hpp
@@ -8,21 +8,21 @@
 
 #pragma once
 
-#include <sycl/__spirv/spirv_types.hpp>       // for Scope, __ocl_event_t
-#include <sycl/access/access.hpp>             // for decorated, mode, addr...
-#include <sycl/detail/common.hpp>             // for NDLoop
-#include <sycl/detail/defines.hpp>            // for __SYCL_TYPE
+#include <sycl/__spirv/spirv_types.hpp> // for Scope, __ocl_event_t
+#include <sycl/access/access.hpp>       // for decorated, mode, addr...
+#include <sycl/detail/assert.hpp>
+#include <sycl/detail/async_work_group_copy_ptr.hpp> // for convertToOpenCLGroupAsyncCopyPtr
+#include <sycl/detail/defines.hpp>                   // for __SYCL_TYPE
 #include <sycl/detail/defines_elementary.hpp> // for __SYCL2020_DEPRECATED
 #include <sycl/detail/fwd/multi_ptr.hpp>
-#include <sycl/detail/generic_type_traits.hpp> // for convertToOpenCLType
-#include <sycl/detail/helpers.hpp>             // for Builder, getSPIRVMemo...
-#include <sycl/detail/type_traits.hpp>         // for is_bool, change_base_...
-#include <sycl/device_event.hpp>               // for device_event
-#include <sycl/id.hpp>                         // for id
-#include <sycl/item.hpp>                       // for item
-#include <sycl/memory_enums.hpp>               // for memory_scope
-#include <sycl/pointers.hpp>                   // for decorated_global_ptr
-#include <sycl/range.hpp>                      // for range
+#include <sycl/detail/helpers.hpp> // for Builder, getSPIRVMemo...
+#include <sycl/detail/nd_loop.hpp> // for NDLoop
+#include <sycl/detail/type_traits/bool_traits.hpp> // for is_bool, change_base_type_t
+#include <sycl/device_event.hpp>                   // for device_event
+#include <sycl/id.hpp>                             // for id
+#include <sycl/memory_enums.hpp>                   // for memory_scope
+#include <sycl/pointers.hpp>                       // for decorated_global_ptr
+#include <sycl/range.hpp>                          // for range
 
 #ifndef __SYCL_DEVICE_ONLY__
 #include <sycl/exception.hpp>
@@ -339,8 +339,9 @@ public:
       const {
 #ifdef __SYCL_DEVICE_ONLY__
     __ocl_event_t E = __spirv_GroupAsyncCopy(
-        __spv::Scope::Workgroup, detail::convertToOpenCLType(dest),
-        detail::convertToOpenCLType(src), numElements, srcStride, 0);
+        __spv::Scope::Workgroup, detail::convertToOpenCLGroupAsyncCopyPtr(dest),
+        detail::convertToOpenCLGroupAsyncCopyPtr(src), numElements, srcStride,
+        0);
     return device_event(E);
 #else
     return nullptr;
@@ -364,8 +365,9 @@ public:
       const {
 #ifdef __SYCL_DEVICE_ONLY__
     __ocl_event_t E = __spirv_GroupAsyncCopy(
-        __spv::Scope::Workgroup, detail::convertToOpenCLType(dest),
-        detail::convertToOpenCLType(src), numElements, destStride, 0);
+        __spv::Scope::Workgroup, detail::convertToOpenCLGroupAsyncCopyPtr(dest),
+        detail::convertToOpenCLGroupAsyncCopyPtr(src), numElements, destStride,
+        0);
     return device_event(E);
 #else
     return nullptr;
@@ -388,8 +390,9 @@ public:
                         [[maybe_unused]] size_t srcStride) const {
 #ifdef __SYCL_DEVICE_ONLY__
     __ocl_event_t E = __spirv_GroupAsyncCopy(
-        __spv::Scope::Workgroup, detail::convertToOpenCLType(dest),
-        detail::convertToOpenCLType(src), numElements, srcStride, 0);
+        __spv::Scope::Workgroup, detail::convertToOpenCLGroupAsyncCopyPtr(dest),
+        detail::convertToOpenCLGroupAsyncCopyPtr(src), numElements, srcStride,
+        0);
     return device_event(E);
 #else
     return nullptr;
@@ -412,8 +415,9 @@ public:
                         [[maybe_unused]] size_t destStride) const {
 #ifdef __SYCL_DEVICE_ONLY__
     __ocl_event_t E = __spirv_GroupAsyncCopy(
-        __spv::Scope::Workgroup, detail::convertToOpenCLType(dest),
-        detail::convertToOpenCLType(src), numElements, destStride, 0);
+        __spv::Scope::Workgroup, detail::convertToOpenCLGroupAsyncCopyPtr(dest),
+        detail::convertToOpenCLGroupAsyncCopyPtr(src), numElements, destStride,
+        0);
     return device_event(E);
 #else
     return nullptr;
