+++ sycl/include/sycl/info/info_desc.hpp
@@ -15,6 +15,7 @@
 // id, range, traits, etc. We have to include some headers before including .def
 // files.
 #include <sycl/aspects.hpp>
+#include <sycl/detail/device_info_types.hpp>
 #include <sycl/detail/type_traits.hpp>
 #include <sycl/ext/oneapi/experimental/device_architecture.hpp>
 #include <sycl/ext/oneapi/experimental/forward_progress.hpp>
