+++ sycl/test/include_deps/sycl_khr_includes_group_algorithms.hpp.cpp
@@ -20,33 +20,39 @@
 // CHECK-NEXT: detail/item_base.hpp
 // CHECK-NEXT: id.hpp
 // CHECK-NEXT: range.hpp
+// CHECK-NEXT: detail/loop.hpp
 // CHECK-NEXT: detail/type_traits.hpp
+// CHECK-NEXT: detail/type_traits/bool_traits.hpp
 // CHECK-NEXT: detail/type_traits/vec_marray_traits.hpp
 // CHECK-NEXT: detail/fwd/multi_ptr.hpp
 // CHECK-NEXT: exception.hpp
 // CHECK-NEXT: detail/string.hpp
 // CHECK-NEXT: functional.hpp
 // CHECK-NEXT: group.hpp
-// CHECK-NEXT: detail/common.hpp
-// CHECK-NEXT: detail/generic_type_traits.hpp
-// CHECK-NEXT: aliases.hpp
-// CHECK-NEXT: bit_cast.hpp
+// CHECK-NEXT: detail/assert.hpp
+// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
 // CHECK-NEXT: detail/fwd/half.hpp
+// CHECK-NEXT: detail/type_traits/integer_traits.hpp
+// CHECK-NEXT: detail/nd_loop.hpp
 // CHECK-NEXT: device_event.hpp
-// CHECK-NEXT: item.hpp
 // CHECK-NEXT: pointers.hpp
 // CHECK-NEXT: half_type.hpp
+// CHECK-NEXT: bit_cast.hpp
 // CHECK-NEXT: aspects.hpp
 // CHECK-NEXT: info/aspects.def
 // CHECK-NEXT: info/aspects_deprecated.def
 // CHECK-NEXT: known_identity.hpp
+// CHECK-NEXT: aliases.hpp
+// CHECK-NEXT: detail/generic_type_traits.hpp
 // CHECK-NEXT: marray.hpp
+// CHECK-NEXT: detail/common.hpp
 // CHECK-NEXT: vector.hpp
 // CHECK-NEXT: detail/named_swizzles_mixin.hpp
 // CHECK-NEXT: detail/vector_arith.hpp
 // CHECK-NEXT: detail/fwd/accessor.hpp
 // CHECK-NEXT: detail/memcpy.hpp
 // CHECK-NEXT: nd_item.hpp
+// CHECK-NEXT: item.hpp
 // CHECK-NEXT: nd_range.hpp
 // CHECK-NEXT: ext/oneapi/functional.hpp
 // CHECK-NEXT: detail/spirv.hpp
