+++ sycl/test/include_deps/sycl_khr_includes_marray.hpp.cpp
@@ -11,7 +11,9 @@
 // CHECK-NEXT: marray.hpp
 // CHECK-NEXT: aliases.hpp
 // CHECK-NEXT: detail/common.hpp
-// CHECK-NEXT: detail/export.hpp
+// CHECK-NEXT: detail/assert.hpp
 // CHECK-NEXT: __spirv/spirv_vars.hpp
+// CHECK-NEXT: detail/export.hpp
+// CHECK-NEXT: detail/nd_loop.hpp
 // CHECK-NEXT: detail/fwd/half.hpp
 // CHECK-EMPTY:
