+++ sycl/test/include_deps/sycl_khr_includes_usm.hpp.cpp
@@ -13,27 +13,29 @@
 // CHECK-NEXT: detail/builtins/builtins.hpp
 // CHECK-NEXT: detail/fwd/multi_ptr.hpp
 // CHECK-NEXT: access/access.hpp
-// CHECK-NEXT: detail/generic_type_traits.hpp
-// CHECK-NEXT: aliases.hpp
-// CHECK-NEXT: bit_cast.hpp
-// CHECK-NEXT: detail/fwd/half.hpp
+// CHECK-NEXT: detail/loop.hpp
 // CHECK-NEXT: detail/type_traits.hpp
+// CHECK-NEXT: detail/type_traits/bool_traits.hpp
 // CHECK-NEXT: detail/type_traits/vec_marray_traits.hpp
-// CHECK-NEXT: detail/helpers.hpp
-// CHECK-NEXT: __spirv/spirv_types.hpp
-// CHECK-NEXT: detail/defines.hpp
-// CHECK-NEXT: detail/export.hpp
-// CHECK-NEXT: memory_enums.hpp
-// CHECK-NEXT: __spirv/spirv_vars.hpp
 // CHECK-NEXT: half_type.hpp
+// CHECK-NEXT: bit_cast.hpp
+// CHECK-NEXT: detail/export.hpp
+// CHECK-NEXT: detail/fwd/half.hpp
 // CHECK-NEXT: aspects.hpp
+// CHECK-NEXT: detail/defines.hpp
 // CHECK-NEXT: info/aspects.def
 // CHECK-NEXT: info/aspects_deprecated.def
 // CHECK-NEXT: marray.hpp
+// CHECK-NEXT: aliases.hpp
 // CHECK-NEXT: detail/common.hpp
+// CHECK-NEXT: detail/assert.hpp
+// CHECK-NEXT: __spirv/spirv_vars.hpp
+// CHECK-NEXT: detail/nd_loop.hpp
 // CHECK-NEXT: vector.hpp
 // CHECK-NEXT: detail/named_swizzles_mixin.hpp
 // CHECK-NEXT: detail/vector_arith.hpp
+// CHECK-NEXT: detail/generic_type_traits.hpp
+// CHECK-NEXT: detail/type_traits/integer_traits.hpp
 // CHECK-NEXT: detail/fwd/accessor.hpp
 // CHECK-NEXT: detail/memcpy.hpp
 // CHECK-NEXT: detail/builtins/common_functions.inc
@@ -52,12 +54,14 @@
 // CHECK-NEXT: range.hpp
 // CHECK-NEXT: info/info_desc.hpp
 // CHECK-NEXT: ur_api.h
+// CHECK-NEXT: detail/device_info_types.hpp
 // CHECK-NEXT: ext/oneapi/experimental/device_architecture.hpp
 // CHECK-NEXT: ext/oneapi/experimental/device_architecture.def
 // CHECK-NEXT: ext/oneapi/experimental/forward_progress.hpp
 // CHECK-NEXT: ext/oneapi/matrix/query-types.hpp
 // CHECK-NEXT: ext/oneapi/bfloat16.hpp
 // CHECK-NEXT: ext/oneapi/matrix/matrix-unified-utils.hpp
+// CHECK-NEXT: __spirv/spirv_types.hpp
 // CHECK-NEXT: info/platform_traits.def
 // CHECK-NEXT: info/context_traits.def
 // CHECK-NEXT: info/device_traits_2020_deprecated.def
@@ -90,6 +94,8 @@
 // CHECK-NEXT: queue.hpp
 // CHECK-NEXT: accessor.hpp
 // CHECK-NEXT: buffer.hpp
+// CHECK-NEXT: detail/helpers.hpp
+// CHECK-NEXT: memory_enums.hpp
 // CHECK-NEXT: detail/iostream_proxy.hpp
 // CHECK-NEXT: detail/is_device_copyable.hpp
 // CHECK-NEXT: detail/stl_type_traits.hpp
@@ -114,9 +120,10 @@
 // CHECK-NEXT: detail/item_base.hpp
 // CHECK-NEXT: detail/kernel_desc.hpp
 // CHECK-NEXT: group.hpp
+// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
 // CHECK-NEXT: device_event.hpp
-// CHECK-NEXT: item.hpp
 // CHECK-NEXT: h_item.hpp
+// CHECK-NEXT: item.hpp
 // CHECK-NEXT: kernel_handler.hpp
 // CHECK-NEXT: nd_item.hpp
 // CHECK-NEXT: nd_range.hpp
